server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
pydantic
pydantic-settings
email-validator
requests
httpx
apify-client
numpy
scikit-learn
apscheduler
google-genai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic
cryptography
redis